+++ library/core/src/main/java/com/google/android/exoplayer2/ExoPlayerImpl.java
@@ -821,6 +821,12 @@ import java.util.concurrent.TimeoutException;
   }
 
   @Override
+  @Nullable
+  public TrackSelector getTrackSelector() {
+    return trackSelector;
+  }
+
+  @Override
   public TrackGroupArray getCurrentTrackGroups() {
     return playbackInfo.trackGroups;
   }