+++ library/core/src/main/java/com/google/android/exoplayer2/SimpleExoPlayer.java
@@ -1558,6 +1558,13 @@ public class SimpleExoPlayer extends BasePlayer
   }
 
   @Override
+  @Nullable
+  public TrackSelector getTrackSelector() {
+    verifyApplicationThread();
+    return player.getTrackSelector();
+  }
+
+  @Override
   public TrackGroupArray getCurrentTrackGroups() {
     verifyApplicationThread();
     return player.getCurrentTrackGroups();